+++ packages/ui/src/index.ts
@@ -47,3 +47,5 @@ export * from './components/ChipPicker';
 export * from './components/SearchInput';
 export * from './components/SegmentedControl';
 export * from './components/AlertBanner';
+export * from './components/TopBar';
+export * from './components/ShortcutHelpModal';